Mass rename and more flexible transient function
The name of the global code structure (e.g. BS, SS, KM, SOL, etc.) have
changed in order to be more explicit.

The desired number of modes that are involved in the modal superposition
is now given as a parameter of the main transient function. This is more
flexible and will enable us to study the convergence of the mode
displacement and mode acceleration method.