Keep the source code constants in a separate file
The constants of the source code, i.e., the default values of the main
file MECA0029, are stored in a separate file, under
`util/load/defaults.m`.

Note that now, the time sample used for time evolution studies is a code
constant, named T_SET. This allows to compute the time evolution in a
more flexible way, by deciding which time sample to use.